Defer replacement of API token placeholders in Caddyfile
Placeholders are now resolved in the underlying JSON config, so it is no longer
necessary to resolve them during Caddyfile unmarshalling.

Possible (but hopefully unlikely) incompatibility: Previously, when a Caddyfile
used a placeholder in the API token, the resulting JSON configuration
(autosaved or exported via the API) would contain the replaced value, i.e. the
actual API token. With this change, the JSON configuration includes the
original placeholder rather than the replaced value.